requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
shap>=0.42.0
lime>=0.2.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Frontend (Streamlit)
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
